requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.94.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
packaging==23.0
pip-review==1.3.0
pydantic==1.10.6
PyJWT==2.6.0
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.6
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
